+++ scripts/cache.js
@@ -19,14 +19,7 @@ catch (error)
 	zstd = zstdEncoder = zstdDecoder = false;
 }
 
-var cacheFolder = p.join(electronRemote.app.getPath('cache'), 'opencomic');
-
-if(process.platform == 'win32' || process.platform == 'win64')
-	cacheFolder = cacheFolder.replace(/AppData\\Roaming/, 'AppData\\Local');
-
-if(!fs.existsSync(cacheFolder)) fs.mkdirSync(cacheFolder);
-cacheFolder = p.join(cacheFolder, 'cache');
-if(!fs.existsSync(cacheFolder)) fs.mkdirSync(cacheFolder);
+var cacheFolder = settings.getCacheFolder();
 
 async function processTheImageQueue(img = false)
 {
@@ -820,7 +813,8 @@ function validate()
 }
 
 module.exports = {
-	folder: cacheFolder,
+	get folder(){return cacheFolder},
+	setCacheFolder: function(folder){cacheFolder = folder},
 	returnThumbnailsImages: returnThumbnailsImages,
 	cleanQueue: cleanQueue,
 	writeFile: writeFile,